EPMEDU-2637-Remove Quick and Nimble dependency
To reduce the number of 3rd party libraries in our app the idea is to remove Quick and Nimble from from project as dependency.
Discussed with Mikhail as of now unit test for this project is not maintained.
And going further any new tests written will need Quick Nimble will be decided later as per use case and latest tech at that point of time.
Note:
As of now we cannot execute unit test from the command line due to the fact we are calling Firebase related API from non main thread, without initialising firebase from main thread. As a part of my improvement effort I will be fixing this issue to ensure all firebase and crashlytics startup from main thread itself before we use it in any form. I expect to raise this pr soon.