Update QuestionMapper.swift (#201)
In the FAQ response some of the values did not have it's corresponding localised strings but had it's non localised counterpart. Hence default to the non localised counterpart. In case the localised part is missing.
source of idea is one of the previous pr's on a similar topic https://github.com/AnimealProject/animeal_iOS/pull/193