EPMEDU-2314: The About Page – Minor Spelling Issues (#195)
* EPMEDU-2314: About page: spelling issues are fixed

* EPMEDU-2314: .gitignore updated

* EPMEDU-2314: after review

* Feature/implement script to generate amplify folder (#192)

* add a script to generate the model folder in amplify folder

added a script to generate the model folder in amplify folder
the script was called from the update amplify script

* update .gitignore

Update .gitignore this change in introduced after running the update amplify script
I had other changes as well but did not commit as it might break for people with lower aws amplify versions
will revert this if needed. assuming this dosen't hurt

* Update update_amplify.sh

call the folders check script a bit later in order
apparently the aws tool creates an amplify folder, which later is converted in to amplify-backup folder
the script is tricked where initially it assumes that the folder is there where as in reality the folder is updated during runtime

* update build script and github actions files

remove the amplify codegen models command from the update amplify script command
make the update update script as fail-able
add another step to recover from error

* Update LoginViewModelTests.swift

update build warning for unit test

* Disable test step for now

disabled test step for now as the test is crashing when running test via command line. logs are attached and explained in detail

* Run the recover from error step only if Amplify setup fails

* Delete check_amplify_generated_folders.sh

Remove un used script

* epmedu-1751: fix resend code for Facebook sign in (#199)

* Fix for resending sms code during Facebook auth

* Minor code fix

* Final fixes for resend code

---------

Co-authored-by: Pran Kishore <kishorepran@gmail.com>
Co-authored-by: Mikhail Churbanov <mihail.churbanov@gmail.com>