EPMEDU-2651 Implement a feature to show a toast message on user copied bank details event
Create a protocol to be used as a delegate to notify the view controller about the event in the view model.
Removed unused methods and protocols from the donate view model
Update the toast view to support showing the view in the bottom of any given view
To provide localisation support. Implement the localisation in the 3 places.
1. Strings.File to update the enum. 2. Sheet.CSV file to keep track of used strings 3. Localisable strings file.