update feeding points selection based on selection updates in search screen
Issue: If the selected feeding points type i.e. cat or dog is updated in search screen. either the update happened via cached data or live data.
update via cached data: no trigger was there to update the home screen feeding point selection
update via live data: there was a trigger to update the home screen feeding point selection but that update the feeding points on the map only and not the selected tab over head.

Implemented the feature as follows:
update via live data: call the method to update the segmented control
update via cached data: call the method on view will appear. it check if the current selection on map screen is same as that of the selection recorded in user defaults. If not then call a method to fetch this cached data and update the selection.
cached data is used for updates in view will appear as we don't want to call network frequently.

Technical changes:
1. Implement a computed property in home model to support fetching filtered cached data.
2. Implement one method in home view model to fetch cached data and trigger update selection in ui + map
3. expose one computed property to check which is the currently selected filter type in user defaults.
4. on view will appear check if we need to update the ui