EPMEDU-2607-Update the profile view model to support discarding the c… (#205)
* EPMEDU-2607-Update the profile view model to support discarding the changes when user clicks no.

Create a new action i.e. discard changes action. it sets the next action as edit action. and updates the profile mutable model to it's initial state.
Add discard action to profile model and use fetch actions to update the next set of action and execute the current action set.
document the profile use cases. as it does more than just fetch profile data. Also multiple generic overloaded methods call call as function, makes it hard to remember which method is doing exactly what.
Finally update the view model to handle the updated back button action.
Update some naming to fix warnings

* update the navigation flow logic for profile view

Update the profile action to return none as the intermediate steps
Update the profile contract to handle the cancel button tap and support did start editing computed property
Set the view controller to show the cancel button in the left to cancel the state if the user started the edit mode
Update the view model to handle the back button action and detect if the edit action is the current action supported