do not auto capitalise text fix major issue on profile screen
in the email text field for the case of registration the user was not able to enter a simple email id without having the first word in capital. Also need to take the user to home screen when the user entered the details during registration and in case of profile update keep the user on the same screen.